+++ src/main/java/com/github/nthykier/debpkg/util/TriConsumer.java
@@ -1,0 +1,5 @@
+package com.github.nthykier.debpkg.util;
+
+public interface TriConsumer<A, B, C> {
+    void accept(A a, B b, C c);
+}